Update README.md
# Silent Words
### Abstract
Sign language is one of the oldest means of communication between humans, and it is one of the languages that must be widely taken care of in order to facilitate communication with deaf and mute people.  <br> Sign language is not limited to the movement of the hands, it includes facial expressions, the movement of the lips and expressions with the movement of the body.
Hence the idea of our site, which aims to teach sign language in the simplest way, which is the **hint:** through the use of hand movements with the use of the language of the lips, and this method of education for the deaf and mute is one of the means that helps to strengthen their spoken language.

### Main page
<img width="1343" alt="Screen Shot 2021-08-27 at 3 08 43 PM" src="https://user-images.githubusercontent.com/76398557/131125383-e87b3925-c5f1-463f-87c9-c2e0f0ae729f.png">
<img width="1343" alt="Screen Shot 2021-08-27 at 3 09 03 PM" src="https://user-images.githubusercontent.com/76398557/131125581-0d9d92f4-83f4-4f61-ad92-8e7307ca18fd.png">

&nbsp; <br>  

## Functionalities:

### 1. Learn Sign Language

Learn sign language online with 200+ vocabulary words divided into many categories that you will use immediately. <br>  

<img width="1343" alt="Screen Shot 2021-08-27 at 4 20 34 PM" src="https://user-images.githubusercontent.com/76398557/131133867-2145d98b-864c-47d7-9bb5-bde35989d484.png">

![category](https://user-images.githubusercontent.com/76398557/131130939-29637547-e574-4222-b93d-b7772938b574.png) 
&nbsp; <br>
&nbsp; <br>  
### 2. Video and text chat

Meet people online and talk to them in sign language ..or ask about any difficulties you face in learning. <br>  

<img width="980" alt="Screen Shot 2021-08-27 at 4 32 48 PM" src="https://user-images.githubusercontent.com/76398557/131135435-6789519f-45cc-4f12-8798-8ee51a779505.png">

![video](https://user-images.githubusercontent.com/76398557/131135265-c383338e-97c4-43f9-a91a-180f27162ee1.png)

### 3. Question bank
Test yourself by using a set of random questions from the question bank and get your result immediately! <br>

![quiz](https://user-images.githubusercontent.com/76398557/131134772-ee36772b-b3a4-4223-904e-9fbe12b1c5dd.png)

<img width="1343" alt="Screen Shot 2021-08-27 at 3 40 28 PM" src="https://user-images.githubusercontent.com/76398557/131134176-0880404d-0d8a-48c0-a230-e5fd99c3b421.png">

### 4. Articles
Read articles that talk about sign language or how to deal with deaf people and many more.. <br> <br>
<img width="1343" alt="Screen Shot 2021-08-27 at 3 31 09 PM" src="https://user-images.githubusercontent.com/76398557/131127711-75aafcd6-b7f6-4843-9fa5-78a9d0c4d8be.png">

&nbsp; <br>  

## Technologies Used:
- MongoDB
- Express
- React
- Node.js
- Material UI
<img width="600" alt=" " src="https://user-images.githubusercontent.com/76398557/131138283-8bb0f392-265a-4ffb-9116-586928857da5.jpg">

&nbsp; <br>  

## Team Members:
- Yousef Swailem (@yousefswailem)
- Hatem Ratrout (@HatemRatrout)
- Mousa Jaffrah (@mousajaffrah)